refactor: apiserver logger
