+++ apiserver/src/lib.rs
@@ -1,3 +1,4 @@
+use logger::LogWriter;
 use nt_analyzer::Analyzer;
 use nt_cmd::config;
 use nt_io::PacketIO;
@@ -5,51 +6,15 @@ use nt_modifier::Modifier;
 use nt_ruleset::expr_rule::ExprRuleset;
 use std::error::Error;
 use std::sync::Arc;
-use tokio::sync::{broadcast, RwLock};
-use tracing_subscriber::fmt::MakeWriter;
+use tokio::sync::RwLock;
 
 pub mod error;
 pub mod file;
+pub mod logger;
 pub mod save;
 pub mod service;
 
 type SharedServerConfig = Arc<RwLock<ServerConfig>>;
-
-#[derive(Clone)]
-pub struct LogWriter {
-    pub tx: broadcast::Sender<String>,
-}
-
-impl LogWriter {
-    pub fn new(capacity: usize) -> Self {
-        let (tx, _rx) = broadcast::channel(capacity);
-        Self { tx }
-    }
-
-    pub async fn subscribe(&self) -> broadcast::Receiver<String> {
-        self.tx.subscribe()
-    }
-}
-
-impl<'a> MakeWriter<'a> for LogWriter {
-    type Writer = LogWriter;
-
-    fn make_writer(&'a self) -> Self::Writer {
-        self.clone()
-    }
-}
-
-impl std::io::Write for LogWriter {
-    fn write(&mut self, buf: &[u8]) -> std::io::Result<usize> {
-        let log = String::from_utf8_lossy(buf).to_string();
-        let _ = self.tx.send(log);
-        Ok(buf.len())
-    }
-
-    fn flush(&mut self) -> std::io::Result<()> {
-        Ok(())
-    }
-}
 
 /// The server configuration.
 pub struct ServerConfig {
